fix unreliable quality dp, add more tests
