+++ docs/index.html
@@ -11,9 +11,9 @@
 
         <!-- config for our rust wasm binary. go to https://trunkrs.dev/assets/#rust for more customization -->
         
-<script type="module" nonce="SDZkyiZFsh9b/onXq5qg8Q==">
-import init, * as bindings from '/raphael-xiv-5cbc2567736698e8.js';
-const wasm = await init('/raphael-xiv-5cbc2567736698e8_bg.wasm');
+<script type="module" nonce="pOIVnbC20b926fLZ61W97A==">
+import init, * as bindings from '/raphael-xiv-d644101ca575d125.js';
+const wasm = await init('/raphael-xiv-d644101ca575d125_bg.wasm');
 
 
 window.wasmBindings = bindings;
@@ -123,8 +123,8 @@ dispatchEvent(new CustomEvent("TrunkApplicationStarted", {detail: {wasm}}));
             }
         </style>
     
-<link rel="modulepreload" href="/raphael-xiv-5cbc2567736698e8.js" crossorigin=anonymous integrity="sha384-yEkE9QIzv/CSWZ+pijZg+H/WRLzgSiuYFPsGl2fExQzqt7ZmaCW4QOE6inFqlWej">
-<link rel="preload" href="/raphael-xiv-5cbc2567736698e8_bg.wasm" crossorigin=anonymous integrity="sha384-L+vDnyxS8nQkCwgq93gqFOHnNwwnrTxtZtNiOooAUGRXefgRSIuau/F3fJQM7J7a" as="fetch" type="application/wasm"></head>
+<link rel="modulepreload" href="/raphael-xiv-d644101ca575d125.js" crossorigin=anonymous integrity="sha384-NNCcNPB4GdK81//Xoj+IbwlJ9JzorzmXxko7F2gProlQfErm64wdub9voi/WqCLy">
+<link rel="preload" href="/raphael-xiv-d644101ca575d125_bg.wasm" crossorigin=anonymous integrity="sha384-UAmz4XoUe/KzrOF1uaOnEhFt7xNdMPcEkiBjljEw+KdNEXCuk3SsIcbCUrfH1byf" as="fetch" type="application/wasm"></head>
 
     <body>
         <!-- The WASM code will resize the canvas dynamically -->
